refactor settings views
