+++ src/features/settings/pages/settings.page.tsx
@@ -1,43 +1,13 @@
+import React from "react";
+import AutoStartSettingsView from "@/features/settings/views/autostart.view";
 import PageWrapper from "@/components/page-wrapper";
 import Card from "@/components/card";
+
 import { Icon24Hours } from "@tabler/icons-react";
-import { Center, Grid, NavLink, Skeleton, Switch } from "@mantine/core";
-import { useEffect, useState } from "react";
-import React from "react";
-import { autostart } from "@/lib";
+import { Center, Grid, NavLink } from "@mantine/core";
+import { useState } from "react";
 
-const AutoStartSettings = () => {
-  const [checked, setChecked] = useState(false);
-  const [loading, setLoading] = useState(true);
-  const checkAutoStart = async () => setChecked(await autostart.isEnabled());
-
-  useEffect(() => {
-    checkAutoStart();
-    setLoading(false);
-  }, [checkAutoStart]);
-
-  const onChange = () => {
-    if (!checked) {
-      autostart.enable();
-    } else {
-      autostart.disable();
-    }
-    setChecked(!checked);
-  };
-
-  if (loading) {
-    return (
-      <>
-        <Skeleton height={8} radius="xl" />
-        <Skeleton height={8} mt={6} radius="xl" />
-        <Skeleton height={8} mt={6} width="70%" radius="xl" />
-      </>
-    );
-  }
-  return <Switch checked={checked} onChange={onChange} label="Start on system startup" />;
-};
-
-const settings = [{ icon: Icon24Hours, label: "Auto Start", view: <AutoStartSettings /> }];
+const settings = [{ icon: Icon24Hours, label: "Auto Start", view: <AutoStartSettingsView /> }];
 
 const SettingsPage = () => {
   const [active, setActive] = useState(0);
